Update SyncUpDetail tutorial code for deletion and dismissal
Refactor SyncUpDetail reducer to:
- Use type-safe `.syncUps` key for shared state
- Implement thread-safe state mutation with `withLock`
