+++ Sources/ComposableArchitecture/Documentation.docc/Tutorials/BuildingSyncUps/06-SyncUpDetail/EditingAndDeletingSyncUp-02-code-0014.swift
@@ -3,111 +3,85 @@ import SwiftUI
 
 @Reducer
 struct SyncUpDetail {
-  // ...
+  @ObservableState
+  struct State: Equatable {
+    @Presents var alert: AlertState<Action.Alert>?
+    @Presents var editSyncUp: SyncUpForm.State?
+    @Shared var syncUp: SyncUp
+  }
+
+  enum Action {
+    case alert(PresentationAction<Alert>)
+    case cancelEditButtonTapped
+    case deleteButtonTapped
+    case doneEditingButtonTapped
+    case editButtonTapped
+    case editSyncUp(PresentationAction<SyncUpForm.Action>)
+    @CasePathable
+    enum Alert {
+      case confirmButtonTapped
+    }
+  }
+
+  @Dependency(\.dismiss) var dismiss
+
+  var body: some ReducerOf<Self> {
+    Reduce { state, action in
+      switch action {
+      case .alert(.presented(.confirmButtonTapped)):
+        @Shared(.syncUps) var syncUps
+        $syncUps.withLock { _ = $0.remove(id: state.syncUp.id) }
+        return .run { _ in await dismiss() }
+
+      case .alert(.dismiss):
+        return .none
+
+      case .cancelEditButtonTapped:
+        state.editSyncUp = nil
+        return .none
+
+      case .deleteButtonTapped:
+        state.alert = .deleteSyncUp
+        return .none
+
+      case .doneEditingButtonTapped:
+        guard let editedSyncUp = state.editSyncUp?.syncUp
+        else { return .none }
+        state.$syncUp.withLock { $0 = editedSyncUp }
+        state.editSyncUp = nil
+        return .none
+
+      case .editButtonTapped:
+        state.editSyncUp = SyncUpForm.State(syncUp: state.syncUp)
+        return .none
+
+      case .editSyncUp:
+        return .none
+      }
+    }
+    .ifLet(\.$editSyncUp, action: \.editSyncUp) {
+      SyncUpForm()
+    }
+    .ifLet(\.$alert, action: \.alert) 
+  }
+}
+
+extension AlertState where Action == SyncUpDetail.Action.Alert {
+  static let deleteSyncUp = Self {
+    TextState("Delete?")
+  } actions: {
+    ButtonState(role: .destructive, action: .confirmButtonTapped) {
+      TextState("Yes")
+    }
+    ButtonState(role: .cancel) {
+      TextState("Nevermind")
+    }
+  } message: {
+    TextState("Are you sure you want to delete this meeting?")
+  }
 }
 
 struct SyncUpDetailView: View {
-  @Bindable var store: StoreOf<SyncUpDetail>
-
-  var body: some View {
-    Form {
-      Section {
-        NavigationLink {
-        } label: {
-          Label("Start Meeting", systemImage: "timer")
-            .font(.headline)
-            .foregroundColor(.accentColor)
-        }
-        HStack {
-          Label("Length", systemImage: "clock")
-          Spacer()
-          Text(store.syncUp.duration.formatted(.units()))
-        }
-
-        HStack {
-          Label("Theme", systemImage: "paintpalette")
-          Spacer()
-          Text(store.syncUp.theme.name)
-            .padding(4)
-            .foregroundColor(store.syncUp.theme.accentColor)
-            .background(store.syncUp.theme.mainColor)
-            .cornerRadius(4)
-        }
-      } header: {
-        Text("Sync-up Info")
-      }
-
-      if !store.syncUp.meetings.isEmpty {
-        Section {
-          ForEach(store.syncUp.meetings) { meeting in
-            Button {
-            } label: {
-              HStack {
-                Image(systemName: "calendar")
-                Text(meeting.date, style: .date)
-                Text(meeting.date, style: .time)
-              }
-            }
-          }
-        } header: {
-          Text("Past meetings")
-        }
-      }
-
-      Section {
-        ForEach(store.syncUp.attendees) { attendee in
-          Label(attendee.name, systemImage: "person")
-        }
-      } header: {
-        Text("Attendees")
-      }
-
-      Section {
-        Button("Delete") {
-          store.send(.deleteButtonTapped)
-        }
-        .foregroundColor(.red)
-        .frame(maxWidth: .infinity)
-      }
-    }
-    .navigationTitle(Text(store.syncUp.title))
-    .toolbar {
-      Button("Edit") {
-        store.send(.editButtonTapped)
-      }
-    }
-    .alert($store.scope(state: \.alert, action: \.alert))
-    .sheet(item: $store.scope(state: \.editSyncUp, action: \.editSyncUp)) { editSyncUpStore in
-      NavigationStack {
-        SyncUpFormView(store: editSyncUpStore)
-          .navigationTitle(store.syncUp.title)
-          .toolbar {
-            ToolbarItem(placement: .cancellationAction) {
-              Button("Cancel") {
-                store.send(.cancelEditButtonTapped)
-              }
-            }
-            ToolbarItem(placement: .confirmationAction) {
-              Button("Done") {
-                store.send(.doneEditingButtonTapped)
-              }
-            }
-          }
-      }
-    }
-  }
+  // ...
 }
 
-#Preview {
-  NavigationStack {
-    SyncUpDetailView(
-      store: Store(
-        initialState: SyncUpDetail.State(
-          syncUp: Shared(value: .mock)
-        )
-      ) {
-        SyncUpDetail()
-      }
-    )
-  }
-}
